KAFKA-17184: Fix the error thrown while looking up offset in RemoteIndexCache
For segments that are uploaded to remote, RemoteIndexCache caches the fetched offset, timestamp, and transaction index entries on the first invocation to remote, then the subsequent invocations are accessed from local.

The remote indexes that are cached locally gets removed on two cases:

1. Remote segments are deleted due to breach by retention size/time and start-offset.
2. The number of cached indexes exceed the remote-log-index-cache size limit of 1 GB (default).

There are two layers of locks used in the RemoteIndexCache. One first-layer lock on the RemoteIndexCache and the second-layer lock on the RemoteIndexCache#Entry.

Issue:

1. The first-layer of lock coordinates the remote-log reader and deleter threads. To ensure that the reader and deleter threads are not blocked on each other, we only take `lock.readLock()` when accessing/deleting a cached index entries.
2. The issue happens when both the reader and deleter threads took the readLock, then the deleter thread marked the index as `markedForCleanup`. Now, the reader thread which holds the `indexEntry` gets an IllegalStateException when accessing it.
3. This is a concurrency issue, where we mark the entry as `markedForCleanup` before removing it from the cache. See RemoteIndexCache#remove, and RemoteIndexCache#removeAll methods.
4. When an entry gets evicted from cache, then the cache remove that entry before calling the evictionListener and all the operations are performed atomically by caffeine cache.

Solution:

1. When the deleter thread marks an Entry for deletion, then we rename the underlying index files with ".deleted" as suffix and add a job to the remote-log-index-cleaner thread which perform the actual cleanup. Previously, the indexes were not accessible once it was marked for deletion. Now, we are allowing to access thoes renamed files until they are removed from disk.
2. Similar to local-log index/segemnt deletion, once the files gets renamed with ".deleted" as suffix then the actual deletion of file happens after file.delete.delay.ms delay of 1 minute.
3. During this time, if the same index entry gets fetched again from remote, then it does not have conflict with the deleted entry as the file names are different.

- Refactor RemoteIndexCache and its tests
- Reduced the fileDeleteDelayMs from 60s to 10s
- updated the unit tests

+++ storage/src/main/java/org/apache/kafka/storage/internals/log/RemoteIndexCache.java
@@ -25,7 +25,7 @@ import org.apache.kafka.server.log.remote.storage.RemoteResourceNotFoundExceptio
 import org.apache.kafka.server.log.remote.storage.RemoteStorageException;
 import org.apache.kafka.server.log.remote.storage.RemoteStorageManager;
 import org.apache.kafka.server.log.remote.storage.RemoteStorageManager.IndexType;
-import org.apache.kafka.server.util.ShutdownableThread;
+import org.apache.kafka.server.util.KafkaScheduler;
 
 import com.github.benmanes.caffeine.cache.Cache;
 import com.github.benmanes.caffeine.cache.Caffeine;
@@ -45,12 +45,11 @@ import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
 import java.util.NoSuchElementException;
-import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.locks.ReadWriteLock;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
 import java.util.function.Function;
 import java.util.stream.Stream;
@@ -92,17 +91,13 @@ public class RemoteIndexCache implements Closeable {
     private final AtomicBoolean isRemoteIndexCacheClosed = new AtomicBoolean(false);
 
     /**
-     * Unbounded queue containing the removed entries from the cache which are waiting to be garbage collected.
-     */
-    private final LinkedBlockingQueue<Entry> expiredIndexes = new LinkedBlockingQueue<>();
-
-    /**
      * Lock used to synchronize close with other read operations. This ensures that when we close, we don't have any other
      * concurrent reads in-progress.
      */
-    private final ReentrantReadWriteLock lock = new ReentrantReadWriteLock();
+    private final ReadWriteLock lock = new ReentrantReadWriteLock();
     private final RemoteStorageManager remoteStorageManager;
-    private final ShutdownableThread cleanerThread;
+    private final KafkaScheduler cleanerScheduler = new KafkaScheduler(1, true, REMOTE_LOG_INDEX_CACHE_CLEANER_THREAD);
+    private int fileDeleteDelayMs = 10_000;
 
     /**
      * Actual cache implementation that this file wraps around.
@@ -130,10 +125,7 @@ public class RemoteIndexCache implements Closeable {
 
         internalCache = initEmptyCache(maxSize);
         init();
-
-        // Start cleaner thread that will clean the expired entries.
-        cleanerThread = createCleanerThread();
-        cleanerThread.start();
+        cleanerScheduler.startup();
     }
 
     public void resizeCacheSize(long remoteLogIndexFileCacheSize) {
@@ -159,15 +151,16 @@ public class RemoteIndexCache implements Closeable {
                 .evictionListener((Uuid key, Entry entry, RemovalCause cause) -> {
                     // Mark the entries for cleanup and add them to the queue to be garbage collected later by the background thread.
                     if (entry != null) {
-                        enqueueEntryForCleanup(entry, key);
+                        enqueueEntryForCleanup(entry);
                     } else {
                         log.error("Received entry as null for key {} when the it is removed from the cache.", key);
                     }
                 }).build();
     }
 
-    public Collection<Entry> expiredIndexes() {
-        return Collections.unmodifiableCollection(expiredIndexes);
+    // Visible for testing
+    public int expiredIdxPendingForDeletion() {
+        return cleanerScheduler.pendingTaskSize();
     }
 
     // Visible for testing
@@ -184,7 +177,7 @@ public class RemoteIndexCache implements Closeable {
         lock.readLock().lock();
         try {
             internalCache.asMap().computeIfPresent(key, (k, v) -> {
-                enqueueEntryForCleanup(v, k);
+                enqueueEntryForCleanup(v);
                 // Returning null to remove the key from the cache
                 return null;
             });
@@ -197,7 +190,9 @@ public class RemoteIndexCache implements Closeable {
         lock.readLock().lock();
         try {
             keys.forEach(key -> internalCache.asMap().computeIfPresent(key, (k, v) -> {
-                enqueueEntryForCleanup(v, k);
+                // Mark then entry for cleanup before removing it from the cache to avoid contention with the
+                // next fetch for the same key.
+                enqueueEntryForCleanup(v);
                 // Returning null to remove the key from the cache
                 return null;
             }));
@@ -206,46 +201,32 @@ public class RemoteIndexCache implements Closeable {
         }
     }
 
-    private void enqueueEntryForCleanup(Entry entry, Uuid key) {
+    private void enqueueEntryForCleanup(Entry entry) {
         try {
             entry.markForCleanup();
-            if (!expiredIndexes.offer(entry)) {
-                log.error("Error while inserting entry {} for key {} into the cleaner queue because queue is full.", entry, key);
-            }
+            Runnable runnable = () -> {
+                try {
+                    entry.cleanup();
+                    log.debug("Cleaned up index entry {}", entry);
+                } catch (Exception ex) {
+                    // do not exit for exceptions other than InterruptedException
+                    log.error("Error occurred while cleaning up expired entry: {}", entry, ex);
+                }
+            };
+            cleanerScheduler.scheduleOnce("delete-index", runnable, fileDeleteDelayMs);
         } catch (IOException e) {
             throw new KafkaException(e);
         }
     }
 
     // Visible for testing
-    public ShutdownableThread cleanerThread() {
-        return cleanerThread;
+    public void setFileDeleteDelayMs(int fileDeleteDelayMs) {
+        this.fileDeleteDelayMs = fileDeleteDelayMs;
     }
 
-    private ShutdownableThread createCleanerThread() {
-        ShutdownableThread thread = new ShutdownableThread(REMOTE_LOG_INDEX_CACHE_CLEANER_THREAD) {
-            public void doWork() {
-                try {
-                    Entry entry = expiredIndexes.take();
-                    log.debug("Cleaning up index entry {}", entry);
-                    entry.cleanup();
-                } catch (InterruptedException ie) {
-                    // cleaner thread should only be interrupted when cache is being closed, else it's an error
-                    if (!isRemoteIndexCacheClosed.get()) {
-                        log.error("Cleaner thread received interruption but remote index cache is not closed", ie);
-                        // propagate the InterruptedException outside to correctly close the thread.
-                        throw new KafkaException(ie);
-                    } else {
-                        log.debug("Cleaner thread was interrupted on cache shutdown");
-                    }
-                } catch (Exception ex) {
-                    // do not exit for exceptions other than InterruptedException
-                    log.error("Error occurred while cleaning up expired entry", ex);
-                }
-            }
-        };
-        thread.setDaemon(true);
-        return thread;
+    // Visible for testing
+    public KafkaScheduler cleanerScheduler() {
+        return cleanerScheduler;
     }
 
     private void init() throws IOException {
@@ -335,7 +316,8 @@ public class RemoteIndexCache implements Closeable {
         log.info("RemoteIndexCache starts up in {} ms.", Time.SYSTEM.hiResClockMs() - start);
     }
 
-    private <T> T loadIndexFile(File file, RemoteLogSegmentMetadata remoteLogSegmentMetadata,
+    private <T> T loadIndexFile(File file,
+                                RemoteLogSegmentMetadata remoteLogSegmentMetadata,
                                 Function<RemoteLogSegmentMetadata, InputStream> fetchRemoteIndex,
                                 Function<File, T> readIndex) throws IOException {
         File indexFile = new File(cacheDir, file.getName());
@@ -358,23 +340,22 @@ public class RemoteIndexCache implements Closeable {
         return index;
     }
 
-    public Entry getIndexEntry(RemoteLogSegmentMetadata remoteLogSegmentMetadata) {
-        if (isRemoteIndexCacheClosed.get()) {
-            throw new IllegalStateException("Unable to fetch index for " +
-                    "segment id=" + remoteLogSegmentMetadata.remoteLogSegmentId().id() + ". Instance is already closed.");
-        }
+    public Entry getIndexEntry(RemoteLogSegmentMetadata metadata) {
+        Uuid uuid = metadata.remoteLogSegmentId().id();
+        throwIfCacheClosed(uuid);
         lock.readLock().lock();
         try {
-            // while this thread was waiting for lock, another thread may have changed the value of isRemoteIndexCacheClosed.
-            // check for index close again
-            if (isRemoteIndexCacheClosed.get()) {
-                throw new IllegalStateException("Unable to fetch index for segment-id = "
-                        + remoteLogSegmentMetadata.remoteLogSegmentId().id() + ". Index instance is already closed.");
-            }
-            return internalCache.get(remoteLogSegmentMetadata.remoteLogSegmentId().id(),
-                    uuid -> createCacheEntry(remoteLogSegmentMetadata));
+            throwIfCacheClosed(uuid);
+            return internalCache.get(uuid, k -> createCacheEntry(metadata));
         } finally {
             lock.readLock().unlock();
+        }
+    }
+
+    private void throwIfCacheClosed(Uuid uuid) {
+        if (isRemoteIndexCacheClosed.get()) {
+            throw new IllegalStateException("Unable to fetch index for segment-id = " + uuid +
+                    ". RemoteIndexCache is closed.");
         }
     }
 
@@ -466,16 +447,12 @@ public class RemoteIndexCache implements Closeable {
             lock.writeLock().lock();
             try {
                 log.info("Close initiated for RemoteIndexCache. Cache stats={}. Cache entries pending delete={}",
-                        internalCache.stats(), expiredIndexes.size());
-                // Initiate shutdown for cleaning thread
-                boolean shutdownRequired = cleanerThread.initiateShutdown();
+                        internalCache.stats(), expiredIdxPendingForDeletion());
+                cleanerScheduler.shutdown();
                 // Close all the opened indexes to force unload mmap memory. This does not delete the index files from disk.
-                internalCache.asMap().forEach((uuid, entry) -> entry.close());
-                // wait for cleaner thread to shutdown
-                if (shutdownRequired) cleanerThread.awaitShutdown();
-
                 // Note that internal cache does not require explicit cleaning/closing. We don't want to invalidate or cleanup
                 // the cache as both would lead to triggering of removal listener.
+                internalCache.asMap().forEach((uuid, entry) -> entry.close());
                 log.info("Close completed for RemoteIndexCache");
             } catch (InterruptedException e) {
                 throw new KafkaException(e);
@@ -554,8 +531,9 @@ public class RemoteIndexCache implements Closeable {
         public OffsetPosition lookupOffset(long targetOffset) {
             entryLock.readLock().lock();
             try {
-                if (markedForCleanup) throw new IllegalStateException("This entry is marked for cleanup");
-                else return offsetIndex.lookup(targetOffset);
+                if (cleanStarted)
+                    throw new IllegalStateException("This entry is marked for cleanup");
+                return offsetIndex.lookup(targetOffset);
             } finally {
                 entryLock.readLock().unlock();
             }
@@ -564,8 +542,8 @@ public class RemoteIndexCache implements Closeable {
         public OffsetPosition lookupTimestamp(long timestamp, long startingOffset) {
             entryLock.readLock().lock();
             try {
-                if (markedForCleanup) throw new IllegalStateException("This entry is marked for cleanup");
-
+                if (cleanStarted)
+                    throw new IllegalStateException("This entry is marked for cleanup");
                 TimestampOffset timestampOffset = timeIndex.lookup(timestamp);
                 return offsetIndex.lookup(Math.max(startingOffset, timestampOffset.offset));
             } finally {
@@ -590,11 +568,9 @@ public class RemoteIndexCache implements Closeable {
         public void cleanup() throws IOException {
             entryLock.writeLock().lock();
             try {
-                markForCleanup();
                 // no-op if clean is done already
-                if (!cleanStarted) {
+                if (!cleanStarted && isMarkedForCleanup()) {
                     cleanStarted = true;
-
                     List<StorageAction<Void, Exception>> actions = List.of(() -> {
                         offsetIndex.deleteIfExists();
                         return null;
@@ -605,7 +581,6 @@ public class RemoteIndexCache implements Closeable {
                         txnIndex.deleteIfExists();
                         return null;
                     });
-
                     tryAll(actions);
                 }
             } finally {
